+++ MRCNN/layer/proposal.py
@@ -72,7 +72,7 @@ class ProposalLayer(KL.Layer):
         scores = inputs[0][:, :, 1]
         # Box deltas [batch, num_rois, 4]
         deltas = inputs[1]
-        deltas = deltas * tf.cast(tf.reshape(self.config.RPN_BBOX_STD_DEV, [1, 1, 4]),tf.float32)
+        deltas = deltas * tf.cast(tf.reshape(self.config.RPN_BBOX_STD_DEV, [1, 1, 4]),tf.float16)
         # Anchors
         anchors = inputs[2]
         proposal_count = inputs[3]
@@ -84,17 +84,17 @@ class ProposalLayer(KL.Layer):
         scores = tf.gather(scores, ix, batch_dims=1, axis=1)
         deltas = tf.gather(deltas, ix, batch_dims=1, axis=1)
         pre_nms_anchors = tf.gather(anchors, ix, batch_dims=1, axis=1)
-        # pre_nms_anchors = tf.cast(tf.vectorized_map(lambda x: denorm_boxes_graph(x,list(self.config.IMAGE_SHAPE)[:2]),pre_nms_anchors), tf.float32)
+        # pre_nms_anchors = tf.cast(tf.vectorized_map(lambda x: denorm_boxes_graph(x,list(self.config.IMAGE_SHAPE)[:2]),pre_nms_anchors), tf.float16)
 
         # Apply deltas to anchors to get refined anchors.
         # [batch, N, (y1, x1, y2, x2)]
         boxes = tf.map_fn(lambda x:apply_box_deltas_graph(*x),
                           [pre_nms_anchors,deltas], 
-                          fn_output_signature=tf.TensorSpec(shape=(self.config.PRE_NMS_LIMIT, 4), dtype=tf.float32))
+                          fn_output_signature=tf.TensorSpec(shape=(self.config.PRE_NMS_LIMIT, 4), dtype=tf.float16))
 
         # Clip to image boundaries. Since we're in normalized coordinates,
         # window = tf.stack([0., 0., *list(self.config.IMAGE_SHAPE-1)[:2]])
-        window = tf.stack([0., 0., 1., 1.])
+        window = tf.cast(tf.stack([0., 0., 1., 1.]), tf.float16)
         boxes = tf.vectorized_map(lambda boxes:clip_boxes_graph(boxes, window),boxes)
 
         # Filter out small boxes
@@ -114,7 +114,7 @@ class ProposalLayer(KL.Layer):
             proposals = tf.pad(proposals, [(0, padding), (0, 0)])
             return proposals
         proposals = tf.map_fn(nms,(boxes, scores),
-                              fn_output_signature=tf.TensorSpec(shape=[None,4], dtype=tf.float32))
+                              fn_output_signature=tf.TensorSpec(shape=[None,4], dtype=tf.float16))
         # proposals = tf.vectorized_map(lambda p: utils.norm_boxes(p, tuple(self.config.IMAGE_SHAPE[:2])), proposals)
         
         return proposals
